chore: Update main.py to include input image name and count
The code changes in main.py include updating the `search` variable to take an input image name and adding a `total_image_count` variable to specify the number of images to be processed. This change improves the flexibility of the program by allowing users to specify the image name and count dynamically.

Please note that the commit message has been generated based on the provided code changes and recent commits.