chore: Update resizeImg.py to resize images in 'Áo Cóm dân tộc thái' folder
This commit updates the resizeImg.py file to resize images in the 'Áo Cóm dân tộc thái' folder. The code changes include modifying the input and output folders, as well as updating the image numbering. This change ensures that all images in the specified folder are resized and standardized.

Please note that the commit message has been generated based on the provided code changes and recent commits.